requirements.txt
pytest
allure-pytest

